+++ ruma-client-api/src/r0/keys/upload_keys.rs
@@ -4,8 +4,9 @@ use std::collections::BTreeMap;
 
 use js_int::UInt;
 use ruma_api::ruma_api;
+use ruma_identifiers::{DeviceKeyAlgorithm, DeviceKeyId};
 
-use super::{AlgorithmAndDeviceId, DeviceKeys, KeyAlgorithm, OneTimeKey};
+use super::{DeviceKeys, OneTimeKey};
 
 ruma_api! {
     metadata: {
@@ -24,13 +25,13 @@ ruma_api! {
 
         /// One-time public keys for "pre-key" messages.
         #[serde(skip_serializing_if = "Option::is_none")]
-        pub one_time_keys: Option<BTreeMap<AlgorithmAndDeviceId, OneTimeKey>>,
+        pub one_time_keys: Option<BTreeMap<DeviceKeyId, OneTimeKey>>,
     }
 
     response: {
         /// For each key algorithm, the number of unclaimed one-time keys of that
         /// type currently held on the server for this device.
-        pub one_time_key_counts: BTreeMap<KeyAlgorithm, UInt>
+        pub one_time_key_counts: BTreeMap<DeviceKeyAlgorithm, UInt>
     }
 
     error: crate::Error