+++ ruma-client-api/src/r0/sync/sync_events.rs
@@ -9,10 +9,10 @@ use ruma_events::{
     presence::PresenceEvent, AnyBasicEvent, AnyStrippedStateEvent, AnySyncEphemeralRoomEvent,
     AnySyncRoomEvent, AnySyncStateEvent, AnyToDeviceEvent,
 };
-use ruma_identifiers::{RoomId, UserId};
+use ruma_identifiers::{DeviceKeyAlgorithm, RoomId, UserId};
 use serde::{Deserialize, Serialize};
 
-use crate::r0::{filter::FilterDefinition, keys::KeyAlgorithm};
+use crate::r0::filter::FilterDefinition;
 
 ruma_api! {
     metadata: {
@@ -87,7 +87,7 @@ ruma_api! {
         /// For each key algorithm, the number of unclaimed one-time keys
         /// currently held on the server for a device.
         #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-        pub device_one_time_keys_count: BTreeMap<KeyAlgorithm, UInt>,
+        pub device_one_time_keys_count: BTreeMap<DeviceKeyAlgorithm, UInt>,
     }
 
     error: crate::Error